+++ py/inst/demoUserScript.py
@@ -1,14 +1,14 @@
-import bussin as bs
+import hal9 as h9
 import pandas as pd
 
 df = pd.DataFrame({'state': ("CA", "WA", "OR"),
     'statistic': (69, 420, 666)})
-bs.set('df', df)
+h9.set('df', df)
 
-bs.dropdown('Dropdown', values = lambda :bs.get('df').columns, on_update=lambda x: bs.set('value', x))
+h9.dropdown('Dropdown', values = lambda :bs.get('df').columns, on_update=lambda x: bs.set('value', x))
 
 def filter_and_show_df(value):
     df = bs.get('df')
     return df[df['state'] == value].to_html()
 
-bs.code('rawhtml', lambda :filter_and_show_df(bs.get('value')))
+h9.code('rawhtml', lambda :filter_and_show_df(bs.get('value')))
