--- py/tests/test_bussin.py
@@ -1,1 +1,0 @@
-from bussin import bussin
