+++ py/tests/test_hal9.py
@@ -1,0 +1,1 @@
+from hal9 import hal9
